+++ src/index.ts
@@ -1,5 +1,8 @@
 export * from "./Component";
+export { default as Attributes } from "./Attributes";
+
 export { default as ComponentStore } from "./ComponentStore";
 export { default as UUID } from "./UUID";
-export * from "./Decorators";
-export * from "./Properties";
+
+import * as AutoEvents from "./experimental/AutoEvents";
+export { AutoEvents as AutoEvents };